refractored a lot and added upload page
