+++ src/main/java/com/example/spotify/controller/SpotifyApiController.java
@@ -5,6 +5,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.UUID;
 
 import java.util.Map;
 
@@ -93,4 +103,25 @@ public class SpotifyApiController {
     public Map<String, DataService.YearlyStats> getTopYears() {
         return dataService.getTopYears();
     }
+
+    @PostMapping("/upload")
+    public ResponseEntity<String> uploadFiles(@RequestParam("files") MultipartFile[] files) {
+        try {
+            String sessionId = UUID.randomUUID().toString(); // or get from cookie
+            Path sessionDir = Paths.get(System.getProperty("java.io.tmpdir"), "spotify", sessionId);
+            Files.createDirectories(sessionDir);
+
+            for (MultipartFile file : files) {
+                Path filePath = sessionDir.resolve(file.getOriginalFilename());
+                Files.write(filePath, file.getBytes());
+            }
+
+            dataService.loadSessionFolder(sessionDir.toString()); // new method you'll add below
+
+            return ResponseEntity.ok(sessionId);
+        } catch (IOException e) {
+            return ResponseEntity.status(500).body("Failed to upload files.");
+        }
+    }
+
 }
